private interface for pregel
