+++ libs/langgraph/langgraph/_internal/__init__.py
@@ -1,0 +1,4 @@
+"""Internal modules for LangGraph.
+
+This module is not part of the public API, and thus stability is not guaranteed.
+"""
