+++ libs/langgraph/langgraph/pregel/_read.py
@@ -13,9 +13,9 @@ from langchain_core.runnables import Runnable, RunnableConfig
 from langgraph._internal._config import merge_configs
 from langgraph._internal._runnable import RunnableCallable, RunnableSeq
 from langgraph.constants import CONF, CONFIG_KEY_READ
+from langgraph.pregel._utils import find_subgraph_pregel
+from langgraph.pregel._write import ChannelWrite
 from langgraph.pregel.protocol import PregelProtocol
-from langgraph.pregel.utils import find_subgraph_pregel
-from langgraph.pregel.write import ChannelWrite
 from langgraph.types import CachePolicy, RetryPolicy
 
 READ_TYPE = Callable[[Union[str, Sequence[str]], bool], Union[Any, dict[str, Any]]]
