+++ libs/langgraph/langgraph/pregel/_runner.py
@@ -33,9 +33,9 @@ from langgraph.constants import (
     TAG_HIDDEN,
 )
 from langgraph.errors import GraphBubbleUp, GraphInterrupt
-from langgraph.pregel.algo import Call
-from langgraph.pregel.executor import Submit
-from langgraph.pregel.retry import arun_with_retry, run_with_retry
+from langgraph.pregel._algo import Call
+from langgraph.pregel._executor import Submit
+from langgraph.pregel._retry import arun_with_retry, run_with_retry
 from langgraph.types import (
     CachePolicy,
     PregelExecutableTask,
