+++ libs/langgraph/langgraph/pregel/_validate.py
@@ -6,7 +6,7 @@ from typing import Any
 from langgraph.channels.base import BaseChannel
 from langgraph.constants import RESERVED
 from langgraph.managed.base import ManagedValueMapping
-from langgraph.pregel.read import PregelNode
+from langgraph.pregel._read import PregelNode
 from langgraph.types import All
