+++ libs/langgraph/langgraph/pregel/_write.py
